+++ feature/src/main/java/com/teamkkumul/feature/meetup/meetupdetail/MeetUpDetailFragment.kt
@@ -5,7 +5,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.teamkkumul.core.ui.base.BindingFragment
 import com.teamkkumul.feature.R
 import com.teamkkumul.feature.databinding.FragmentMeetUpDetailBinding
-import com.teamkkumul.feature.mygroup.MyGroupMeetUpItemDecoration
+import com.teamkkumul.feature.utils.itemdecorator.MeetUpFriendItemDecoration
 import com.teamkkumul.model.MeetUpSealedItem
 
 class MeetUpDetailFragment :
@@ -35,7 +35,7 @@ class MeetUpDetailFragment :
             layoutManager =
                 LinearLayoutManager(requireContext(), LinearLayoutManager.HORIZONTAL, false)
             adapter = memberAdapter
-            addItemDecoration(MyGroupMeetUpItemDecoration(requireContext()))
+            addItemDecoration(MeetUpFriendItemDecoration(requireContext()))
         }
     }
