+++ feature/src/main/java/com/teamkkumul/feature/mygroup/mygroupdetail/viewholder/MyGroupDetailMeetUpViewHolder.kt
@@ -1,4 +1,4 @@
-package com.teamkkumul.feature.mygroup.viewholder
+package com.teamkkumul.feature.mygroup.mygroupdetail.viewholder
 
 import android.view.View
 import androidx.core.content.ContextCompat
